Remove dispatch_compat*
Everything is now in one of these files:
  - dispatch.{c,h}
  - php{5,7}/dispatch.c
  - php{5,7}/engine_hooks.c
