+++ package.xml
@@ -53,10 +53,6 @@
                     <file name="vendor_stdatomic.h" role="src" />
                     <file name="dispatch.c" role="src" />
                     <file name="dispatch.h" role="src" />
-                    <file name="dispatch_compat.h" role="src" />
-                    <file name="dispatch_compat_php5.c" role="src" />
-                    <file name="dispatch_compat_php5.h" role="src" />
-                    <file name="dispatch_compat_php7.c" role="src" />
                     <file name="dispatch_setup.c" role="src" />
                     <file name="engine_hooks.c" role="src" />
                     <file name="engine_hooks.h" role="src" />
@@ -86,9 +82,11 @@
                         <file name="mpack.h" role="src" />
                     </dir>
                     <dir name="php5">
+                        <file name="dispatch.c" role="src" />
                         <file name="engine_hooks.c" role="src" />
                     </dir>
                     <dir name="php7">
+                        <file name="dispatch.c" role="src" />
                         <file name="engine_hooks.c" role="src" />
                     </dir>
                     <dir name="third-party">
