+++ src/ext/dispatch.h
@@ -25,4 +25,50 @@ void ddtrace_dispatch_init(TSRMLS_D);
 void ddtrace_dispatch_destroy(TSRMLS_D);
 void ddtrace_dispatch_reset(TSRMLS_D);
 
+#if PHP_VERSION_ID < 70000
+
+#undef EX
+#define EX(element) ((execute_data)->element)
+
+#if PHP_VERSION_ID < 50500
+#define FBC() EX(fbc)
+#define NUM_ADDITIONAL_ARGS() (0)
+#define OBJECT() EX(object)
+#elif PHP_VERSION_ID < 50600
+#define FBC() (EX(call)->fbc)
+#define NUM_ADDITIONAL_ARGS() (0)
+#define OBJECT() (EX(call) ? EX(call)->object : NULL)
+#else
+#define FBC() (EX(call)->fbc)
+#define NUM_ADDITIONAL_ARGS() EX(call)->num_additional_args
+#define OBJECT() (EX(call) ? EX(call)->object : NULL)
+#endif
+
+inline void *zend_hash_str_find_ptr(const HashTable *ht, const char *key, size_t length) {
+    void **rv = NULL;
+    zend_hash_find(ht, key, length, (void **)&rv);
+
+    if (rv) {
+        return *rv;
+    } else {
+        return NULL;
+    }
+}
+
+void ddtrace_class_lookup_release_compat(void *zv);
+
+#define ddtrace_zval_ptr_dtor(x) zval_dtor(x)
+#else
+void ddtrace_class_lookup_release_compat(zval *zv);
+
+#define ddtrace_zval_ptr_dtor(x) zval_ptr_dtor(x)
+#define INIT_ZVAL(x) ZVAL_NULL(&x)
+#endif
+
+zend_function *ddtrace_function_get(const HashTable *table, zval *name);
+void ddtrace_dispatch_free_owned_data(ddtrace_dispatch_t *dispatch);
+HashTable *ddtrace_new_class_lookup(zval *clazz TSRMLS_DC);
+zend_bool ddtrace_dispatch_store(HashTable *class_lookup, ddtrace_dispatch_t *dispatch);
+void ddtrace_wrapper_forward_call_from_userland(zend_execute_data *execute_data, zval *return_value TSRMLS_DC);
+
 #endif  // DISPATCH_H
