+++ src/ext/dispatch_setup.c
@@ -9,9 +9,8 @@
 
 #include "compat_string.h"
 #include "ddtrace.h"
-#include "debug.h"
 #include "dispatch.h"
-#include "dispatch_compat.h"
+
 ZEND_EXTERN_MODULE_GLOBALS(ddtrace);
 
 #if PHP_VERSION_ID >= 70000
