+++ src/ext/php5/engine_hooks.c
@@ -11,8 +11,6 @@
 #include "ddtrace.h"
 #include "debug.h"
 #include "dispatch.h"
-#include "dispatch_compat.h"
-#include "dispatch_compat_php5.h"
 #include "logging.h"
 
 ZEND_EXTERN_MODULE_GLOBALS(ddtrace);
@@ -26,6 +24,257 @@ static user_opcode_handler_t _prev_fcall_by_name_handler;
 #define CTOR_CALL_BIT 0x1
 #define CTOR_USED_BIT 0x2
 #define DECODE_CTOR(ce) ((zend_class_entry *)(((zend_uintptr_t)(ce)) & ~(CTOR_CALL_BIT | CTOR_USED_BIT)))
+
+static zval *ddtrace_this(zend_execute_data *execute_data) {
+    zval *this = NULL;
+#if PHP_VERSION_ID < 50500
+    if (EX(opline)->opcode != ZEND_DO_FCALL && EX(object)) {
+        this = EX(object);
+    }
+#else
+    if (EX(opline)->opcode != ZEND_DO_FCALL) {
+        this = EX(call) ? EX(call)->object : NULL;
+    }
+#endif
+    if (this && Z_TYPE_P(this) != IS_OBJECT) {
+        this = NULL;
+    }
+
+    return this;
+}
+
+static void **vm_stack_push_args_with_copy(int count TSRMLS_DC) {
+    zend_vm_stack p = EG(argument_stack);
+
+    zend_vm_stack_extend(count + 1 TSRMLS_CC);
+
+    EG(argument_stack)->top += count;
+    *(EG(argument_stack)->top) = (void *)(zend_uintptr_t)count;
+    while (count-- > 0) {
+        void *data = *(--p->top);
+
+        if (UNEXPECTED(p->top == ZEND_VM_STACK_ELEMETS(p))) {
+            zend_vm_stack r = p;
+
+            EG(argument_stack)->prev = p->prev;
+            p = p->prev;
+            efree(r);
+        }
+        *(ZEND_VM_STACK_ELEMETS(EG(argument_stack)) + count) = data;
+    }
+    return EG(argument_stack)->top++;
+}
+
+static void **vm_stack_push_args(int count TSRMLS_DC) {
+    if (UNEXPECTED(EG(argument_stack)->top - ZEND_VM_STACK_ELEMETS(EG(argument_stack)) < count) ||
+        UNEXPECTED(EG(argument_stack)->top == EG(argument_stack)->end)) {
+        return vm_stack_push_args_with_copy(count TSRMLS_CC);
+    }
+    *(EG(argument_stack)->top) = (void *)(zend_uintptr_t)count;
+    return EG(argument_stack)->top++;
+}
+
+static void setup_fcal_name(zend_execute_data *execute_data, zend_fcall_info *fci, zval **result TSRMLS_DC) {
+    int argc = EX(opline)->extended_value + NUM_ADDITIONAL_ARGS();
+    fci->param_count = argc;
+
+    if (NUM_ADDITIONAL_ARGS()) {
+        vm_stack_push_args(fci->param_count TSRMLS_CC);
+    } else {
+        if (fci->param_count) {
+            EX(function_state).arguments = zend_vm_stack_top(TSRMLS_C);
+        }
+        zend_vm_stack_push((void *)(zend_uintptr_t)fci->param_count TSRMLS_CC);
+    }
+
+    if (fci->param_count) {
+        fci->params = (zval ***)safe_emalloc(sizeof(zval *), fci->param_count, 0);
+        zend_get_parameters_array_ex(fci->param_count, fci->params);
+    }
+#if PHP_VERSION_ID < 50500
+    if (EG(return_value_ptr_ptr)) {
+        fci->retval_ptr_ptr = EG(return_value_ptr_ptr);
+    } else {
+        fci->retval_ptr_ptr = result;
+    }
+#else
+    fci->retval_ptr_ptr = result;
+#endif
+}
+
+static void ddtrace_setup_fcall(zend_execute_data *execute_data, zend_fcall_info *fci, zval **result TSRMLS_DC) {
+    if (EX(opline)->opcode != ZEND_DO_FCALL_BY_NAME) {
+#if PHP_VERSION_ID >= 50600
+        call_slot *call = EX(call_slots) + EX(opline)->op2.num;
+        call->fbc = EX(function_state).function;
+        call->object = NULL;
+        call->called_scope = NULL;
+        call->num_additional_args = 0;
+        call->is_ctor_call = 0;
+        EX(call) = call;
+#else
+        FBC() = EX(function_state).function;
+#endif
+    }
+
+#if PHP_VERSION_ID < 50500
+    EX(original_return_value) = EG(return_value_ptr_ptr);
+    EG(return_value_ptr_ptr) = result;
+#endif
+
+    setup_fcal_name(execute_data, fci, result TSRMLS_CC);
+}
+
+static zend_function *_get_current_fbc(zend_execute_data *execute_data TSRMLS_DC) {
+    if (EX(opline)->opcode == ZEND_DO_FCALL_BY_NAME) {
+        return FBC();
+    }
+    zend_op *opline = EX(opline);
+    zend_function *fbc = NULL;
+    zval *fname = opline->op1.zv;
+
+    if (CACHED_PTR(opline->op1.literal->cache_slot)) {
+        return CACHED_PTR(opline->op1.literal->cache_slot);
+    } else if (EXPECTED(zend_hash_quick_find(EG(function_table), Z_STRVAL_P(fname), Z_STRLEN_P(fname) + 1,
+                                             Z_HASH_P(fname), (void **)&fbc) == SUCCESS)) {
+        return fbc;
+    } else {
+        return NULL;
+    }
+}
+
+static BOOL_T ddtrace_should_trace_call(zend_execute_data *execute_data, zend_function **fbc,
+                                        ddtrace_dispatch_t **dispatch TSRMLS_DC) {
+    if (DDTRACE_G(disable) || DDTRACE_G(disable_in_current_request) || DDTRACE_G(class_lookup) == NULL ||
+        DDTRACE_G(function_lookup) == NULL) {
+        return FALSE;
+    }
+    *fbc = _get_current_fbc(execute_data TSRMLS_CC);
+    if (!*fbc) {
+        return FALSE;
+    }
+
+    zval zv, *fname;
+    fname = &zv;
+    if (EX(opline)->opcode == ZEND_DO_FCALL_BY_NAME) {
+        ZVAL_STRING(fname, (*fbc)->common.function_name, 0);
+    } else if (EX(opline)->op1.zv) {
+        fname = EX(opline)->op1.zv;
+    } else {
+        return FALSE;
+    }
+
+    // Don't trace closures
+    if ((*fbc)->common.fn_flags & ZEND_ACC_CLOSURE) {
+        return FALSE;
+    }
+
+    zval *this = ddtrace_this(execute_data);
+    *dispatch = ddtrace_find_dispatch(this, *fbc, fname TSRMLS_CC);
+    if (!*dispatch || (*dispatch)->busy) {
+        return FALSE;
+    }
+
+    return TRUE;
+}
+
+int ddtrace_forward_call(zend_execute_data *execute_data, zend_function *fbc, zval *return_value TSRMLS_DC) {
+    int fcall_status;
+
+    zend_fcall_info fci = {0};
+    zend_fcall_info_cache fcc = {0};
+    zval *retval_ptr = NULL;
+
+    fcc.initialized = 1;
+    fcc.function_handler = fbc;
+    fcc.object_ptr = ddtrace_this(execute_data);
+    fcc.calling_scope = fbc->common.scope;  // EG(scope);
+#if PHP_VERSION_ID < 50500
+    fcc.called_scope = EX(called_scope);
+#else
+    fcc.called_scope = EX(call) ? EX(call)->called_scope : NULL;
+#endif
+
+    ddtrace_setup_fcall(execute_data, &fci, &retval_ptr TSRMLS_CC);
+    fci.size = sizeof(fci);
+    fci.no_separation = 1;
+    fci.object_ptr = fcc.object_ptr;
+
+    fcall_status = zend_call_function(&fci, &fcc TSRMLS_CC);
+    if (fcall_status == SUCCESS && fci.retval_ptr_ptr && *fci.retval_ptr_ptr) {
+        COPY_PZVAL_TO_ZVAL(*return_value, *fci.retval_ptr_ptr);
+    }
+
+    zend_fcall_info_args_clear(&fci, 1);
+    return fcall_status;
+}
+
+BOOL_T ddtrace_execute_tracing_closure(zval *callable, zval *span_data, zend_execute_data *execute_data,
+                                       zval *user_args, zval *user_retval, zval *exception TSRMLS_DC) {
+    BOOL_T status = TRUE;
+    zend_fcall_info fci = {0};
+    zend_fcall_info_cache fcc = {0};
+    zval *retval_ptr = NULL;
+    zval **args[4];
+    zval *null_zval = &EG(uninitialized_zval);
+    zval *this = ddtrace_this(execute_data);
+
+    if (zend_fcall_info_init(callable, 0, &fci, &fcc, NULL, NULL TSRMLS_CC) == FAILURE) {
+        ddtrace_log_debug("Could not init tracing closure");
+        return FALSE;
+    }
+
+    /* Note: In PHP 5 there is a bug where closures are automatically
+     * marked as static if they are defined from a static method context.
+     * @see https://3v4l.org/Rgo87
+     */
+    if (this) {
+        BOOL_T is_instance_method = (FBC()->common.fn_flags & ZEND_ACC_STATIC) ? FALSE : TRUE;
+        BOOL_T is_closure_static = (fcc.function_handler->common.fn_flags & ZEND_ACC_STATIC) ? TRUE : FALSE;
+        if (is_instance_method && is_closure_static) {
+            ddtrace_log_debug("Cannot trace non-static method with static tracing closure");
+            return FALSE;
+        }
+    }
+
+    // Arg 0: DDTrace\SpanData $span
+    args[0] = &span_data;
+
+    // Arg 1: array $args
+    args[1] = &user_args;
+
+    // Arg 2: mixed $retval
+    args[2] = &user_retval;
+    // Arg 3: Exception|null $exception
+    args[3] = exception ? &exception : &null_zval;
+
+    fci.param_count = 4;
+    fci.params = args;
+    fci.retval_ptr_ptr = &retval_ptr;
+
+    fcc.initialized = 1;
+    fcc.object_ptr = this;
+#if PHP_VERSION_ID < 50500
+    fcc.called_scope = EX(called_scope);
+#else
+    fcc.called_scope = EX(call) ? EX(call)->called_scope : NULL;
+#endif
+    // Give the tracing closure access to private & protected class members
+    fcc.function_handler->common.scope = fcc.called_scope;
+
+    if (zend_call_function(&fci, &fcc TSRMLS_CC) == FAILURE) {
+        ddtrace_log_debug("Could not execute tracing closure");
+    }
+
+    if (fci.retval_ptr_ptr && retval_ptr) {
+        if (Z_TYPE_P(retval_ptr) == IS_BOOL) {
+            status = Z_LVAL_P(retval_ptr) ? TRUE : FALSE;
+        }
+        zval_ptr_dtor(&retval_ptr);
+    }
+    zend_fcall_info_args_clear(&fci, 0);
+    return status;
+}
 
 static void update_opcode_leave(zend_execute_data *execute_data TSRMLS_DC) {
     DD_PRINTF("Update opcode leave");
@@ -49,6 +298,14 @@ static void update_opcode_leave(zend_execute_data *execute_data TSRMLS_DC) {
     zend_vm_stack_clear_multiple(0 TSRMLS_CC);
     EX(call)--;
 #endif
+}
+
+static zend_function *datadog_current_function(zend_execute_data *execute_data) {
+    if (EX(opline)->opcode == ZEND_DO_FCALL_BY_NAME) {
+        return FBC();
+    } else {
+        return EX(function_state).function;
+    }
 }
 
 static void execute_fcall(ddtrace_dispatch_t *dispatch, zval *this, zend_execute_data *execute_data,
@@ -129,7 +386,7 @@ _exit_cleanup:
     DDTRACE_G(original_context).fbc = current_fbc;
 }
 
-static zend_always_inline void wrap_and_run(zend_execute_data *execute_data, ddtrace_dispatch_t *dispatch TSRMLS_DC) {
+static void wrap_and_run(zend_execute_data *execute_data, ddtrace_dispatch_t *dispatch TSRMLS_DC) {
     zval *this = ddtrace_this(execute_data);
 
 #if PHP_VERSION_ID < 50500
@@ -258,8 +515,8 @@ static void ddtrace_span_attach_exception(ddtrace_span_t *span, ddtrace_exceptio
     }
 }
 
-void ddtrace_trace_dispatch(ddtrace_dispatch_t *dispatch, zend_function *fbc,
-                            zend_execute_data *execute_data TSRMLS_DC) {
+static void ddtrace_trace_dispatch(ddtrace_dispatch_t *dispatch, zend_function *fbc,
+                                   zend_execute_data *execute_data TSRMLS_DC) {
     int fcall_status;
     const zend_op *opline = EX(opline);
 
@@ -344,6 +601,26 @@ void ddtrace_trace_dispatch(ddtrace_dispatch_t *dispatch, zend_function *fbc,
 #endif
 }
 
+static int ddtrace_opcode_default_dispatch(zend_execute_data *execute_data TSRMLS_DC) {
+    if (!EX(opline)->opcode) {
+        return ZEND_USER_OPCODE_DISPATCH;
+    }
+    switch (EX(opline)->opcode) {
+        case ZEND_DO_FCALL:
+            if (_prev_fcall_handler) {
+                return _prev_fcall_handler(execute_data TSRMLS_CC);
+            }
+            break;
+
+        case ZEND_DO_FCALL_BY_NAME:
+            if (_prev_fcall_by_name_handler) {
+                return _prev_fcall_by_name_handler(execute_data TSRMLS_CC);
+            }
+            break;
+    }
+    return ZEND_USER_OPCODE_DISPATCH;
+}
+
 int ddtrace_wrap_fcall(zend_execute_data *execute_data TSRMLS_DC) {
     zend_function *current_fbc = NULL;
     ddtrace_dispatch_t *dispatch = NULL;
@@ -399,24 +676,4 @@ void ddtrace_opcode_minit(void) {
 void ddtrace_opcode_mshutdown(void) {
     zend_set_user_opcode_handler(ZEND_DO_FCALL, NULL);
     zend_set_user_opcode_handler(ZEND_DO_FCALL_BY_NAME, NULL);
-}
-
-int ddtrace_opcode_default_dispatch(zend_execute_data *execute_data TSRMLS_DC) {
-    if (!EX(opline)->opcode) {
-        return ZEND_USER_OPCODE_DISPATCH;
-    }
-    switch (EX(opline)->opcode) {
-        case ZEND_DO_FCALL:
-            if (_prev_fcall_handler) {
-                return _prev_fcall_handler(execute_data TSRMLS_CC);
-            }
-            break;
-
-        case ZEND_DO_FCALL_BY_NAME:
-            if (_prev_fcall_by_name_handler) {
-                return _prev_fcall_by_name_handler(execute_data TSRMLS_CC);
-            }
-            break;
-    }
-    return ZEND_USER_OPCODE_DISPATCH;
 }
