+++ src/ext/php7/engine_hooks.c
@@ -9,7 +9,6 @@
 #include "ddtrace.h"
 #include "debug.h"
 #include "dispatch.h"
-#include "dispatch_compat.h"
 #include "env_config.h"
 #include "logging.h"
 
@@ -27,7 +26,24 @@ static user_opcode_handler_t _prev_fcall_by_name_handler;
 #define RETURN_VALUE_USED(opline) ((opline)->result_type != IS_UNUSED)
 #endif
 
-BOOL_T ddtrace_should_trace_call(zend_execute_data *execute_data, zend_function **fbc, ddtrace_dispatch_t **dispatch) {
+static zval *ddtrace_this(zend_execute_data *execute_data) {
+    // todo: simplify this, and maybe expect it to pass callee e.g. EX(call) instead of caller e.g. execute_data
+    zval *this = NULL;
+
+    if (EX(call)) {
+        if (Z_OBJ(EX(call)->This) != NULL) {
+            this = &(EX(call)->This);
+        }
+    }
+    if (this && Z_TYPE_P(this) != IS_OBJECT) {
+        this = NULL;
+    }
+
+    return this;
+}
+
+static BOOL_T ddtrace_should_trace_call(zend_execute_data *execute_data, zend_function **fbc,
+                                        ddtrace_dispatch_t **dispatch) {
     if (DDTRACE_G(disable) || DDTRACE_G(disable_in_current_request) || DDTRACE_G(class_lookup) == NULL ||
         DDTRACE_G(function_lookup) == NULL) {
         return FALSE;
@@ -99,8 +115,110 @@ static void ddtrace_span_attach_exception(ddtrace_span_t *span, ddtrace_exceptio
     }
 }
 
-void ddtrace_trace_dispatch(ddtrace_dispatch_t *dispatch, zend_function *fbc,
-                            zend_execute_data *execute_data TSRMLS_DC) {
+static void ddtrace_setup_fcall(zend_execute_data *execute_data, zend_fcall_info *fci, zval **result) {
+    fci->param_count = ZEND_CALL_NUM_ARGS(execute_data);
+    fci->params = fci->param_count ? ZEND_CALL_ARG(execute_data, 1) : NULL;
+    fci->retval = *result;
+}
+
+static int ddtrace_forward_call(zend_execute_data *execute_data, zend_function *fbc, zval *return_value,
+                                zend_fcall_info *fci, zend_fcall_info_cache *fcc) {
+    int fcall_status;
+
+#if PHP_VERSION_ID < 70300
+    fcc->initialized = 1;
+#endif
+    fcc->function_handler = fbc;
+    fcc->object = Z_TYPE(EX(This)) == IS_OBJECT ? Z_OBJ(EX(This)) : NULL;
+    fcc->calling_scope = fbc->common.scope;  // EG(scope);
+    fcc->called_scope = zend_get_called_scope(execute_data);
+
+    fci->size = sizeof(zend_fcall_info);
+    fci->no_separation = 1;
+    fci->object = fcc->object;
+
+    ddtrace_setup_fcall(execute_data, fci, &return_value);
+
+    fcall_status = zend_call_function(fci, fcc);
+    if (fcall_status == SUCCESS && Z_TYPE_P(return_value) != IS_UNDEF) {
+#if PHP_VERSION_ID >= 70100
+        if (Z_ISREF_P(return_value)) {
+            zend_unwrap_reference(return_value);
+        }
+#endif
+    }
+
+    // We don't want to clear the args with zend_fcall_info_args_clear() yet
+    // since our tracing closure might need them
+    return fcall_status;
+}
+
+static BOOL_T ddtrace_execute_tracing_closure(zval *callable, zval *span_data, zend_execute_data *execute_data,
+                                              zval *user_args, zval *user_retval, zend_object *exception) {
+    BOOL_T status = TRUE;
+    zend_fcall_info fci = {0};
+    zend_fcall_info_cache fcc = {0};
+    zval rv;
+    INIT_ZVAL(rv);
+    zval args[4];
+    zval exception_arg = {.value = {0}};
+    ZVAL_UNDEF(&exception_arg);
+    if (exception) {
+        ZVAL_OBJ(&exception_arg, exception);
+    }
+    zval *this = ddtrace_this(execute_data);
+
+    if (zend_fcall_info_init(callable, 0, &fci, &fcc, NULL, NULL) == FAILURE) {
+        ddtrace_log_debug("Could not init tracing closure");
+        return FALSE;
+    }
+
+    if (this) {
+        BOOL_T is_instance_method = (EX(call)->func->common.fn_flags & ZEND_ACC_STATIC) ? FALSE : TRUE;
+        BOOL_T is_closure_static = (fcc.function_handler->common.fn_flags & ZEND_ACC_STATIC) ? TRUE : FALSE;
+        if (is_instance_method && is_closure_static) {
+            ddtrace_log_debug("Cannot trace non-static method with static tracing closure");
+            return FALSE;
+        }
+    }
+
+    // Arg 0: DDTrace\SpanData $span
+    ZVAL_COPY(&args[0], span_data);
+
+    // Arg 1: array $args
+    ZVAL_COPY(&args[1], user_args);
+
+    // Arg 2: mixed $retval
+    ZVAL_COPY(&args[2], user_retval);
+
+    // Arg 3: Exception|null $exception
+    ZVAL_COPY(&args[3], &exception_arg);
+
+    fci.param_count = 4;
+    fci.params = args;
+    fci.retval = &rv;
+
+#if PHP_VERSION_ID < 70300
+    fcc.initialized = 1;
+#endif
+    fcc.object = this ? Z_OBJ_P(this) : NULL;
+    fcc.called_scope = zend_get_called_scope(EX(call));
+    // Give the tracing closure access to private & protected class members
+    fcc.function_handler->common.scope = fcc.called_scope;
+
+    if (zend_call_function(&fci, &fcc) == FAILURE) {
+        ddtrace_log_debug("Could not execute tracing closure");
+        status = FALSE;
+    } else if (Z_TYPE(rv) == IS_FALSE) {
+        status = FALSE;
+    }
+
+    zval_ptr_dtor(&rv);
+    zend_fcall_info_args_clear(&fci, 0);
+    return status;
+}
+
+static void ddtrace_trace_dispatch(ddtrace_dispatch_t *dispatch, zend_function *fbc, zend_execute_data *execute_data) {
     int fcall_status;
     const zend_op *opline = EX(opline);
 
@@ -112,10 +230,10 @@ void ddtrace_trace_dispatch(ddtrace_dispatch_t *dispatch, zend_function *fbc,
     user_args = &user_args_zv;
     user_retval = (RETURN_VALUE_USED(opline) ? EX_VAR(opline->result.var) : &rv);
 
-    ddtrace_span_t *span = ddtrace_open_span(TSRMLS_C);
+    ddtrace_span_t *span = ddtrace_open_span();
     zend_fcall_info fci = {0};
     zend_fcall_info_cache fcc = {0};
-    fcall_status = ddtrace_forward_call(EX(call), fbc, user_retval, &fci, &fcc TSRMLS_CC);
+    fcall_status = ddtrace_forward_call(EX(call), fbc, user_retval, &fci, &fcc);
     dd_trace_stop_span_time(span);
 
     ddtrace_copy_function_args(EX(call), user_args);
@@ -125,7 +243,7 @@ void ddtrace_trace_dispatch(ddtrace_dispatch_t *dispatch, zend_function *fbc,
         prev_exception = EG(prev_exception);
         EG(prev_exception) = NULL;
         ddtrace_span_attach_exception(span, exception);
-        zend_clear_exception(TSRMLS_C);
+        zend_clear_exception();
     }
 
     BOOL_T keep_span = TRUE;
@@ -133,16 +251,16 @@ void ddtrace_trace_dispatch(ddtrace_dispatch_t *dispatch, zend_function *fbc,
         zend_error_handling error_handling;
         int orig_error_reporting = EG(error_reporting);
         EG(error_reporting) = 0;
-        zend_replace_error_handling(EH_THROW, NULL, &error_handling TSRMLS_CC);
+        zend_replace_error_handling(EH_THROW, NULL, &error_handling);
         keep_span = ddtrace_execute_tracing_closure(&dispatch->callable, span->span_data, execute_data, user_args,
-                                                    user_retval, exception TSRMLS_CC);
-        zend_restore_error_handling(&error_handling TSRMLS_CC);
+                                                    user_retval, exception);
+        zend_restore_error_handling(&error_handling);
         EG(error_reporting) = orig_error_reporting;
         // If the tracing closure threw an exception, ignore it to not impact the original call
         if (EG(exception)) {
             ddtrace_log_debug("Exeception thrown in the tracing closure");
             if (!DDTRACE_G(strict_mode)) {
-                zend_clear_exception(TSRMLS_C);
+                zend_clear_exception();
             }
         }
     }
@@ -150,16 +268,16 @@ void ddtrace_trace_dispatch(ddtrace_dispatch_t *dispatch, zend_function *fbc,
     zval_ptr_dtor(user_args);
 
     if (keep_span == TRUE) {
-        ddtrace_close_span(TSRMLS_C);
+        ddtrace_close_span();
     } else {
-        ddtrace_drop_span(TSRMLS_C);
+        ddtrace_drop_span();
     }
 
     if (exception) {
         EG(exception) = exception;
         EG(prev_exception) = prev_exception;
 
-        zend_throw_exception_internal(NULL TSRMLS_CC);
+        zend_throw_exception_internal(NULL);
     }
 
     zend_fcall_info_args_clear(&fci, 0);
@@ -176,12 +294,13 @@ void ddtrace_trace_dispatch(ddtrace_dispatch_t *dispatch, zend_function *fbc,
     EX(call) = EX(call)->prev_execute_data;
 }
 
-static void update_opcode_leave(zend_execute_data *execute_data TSRMLS_DC) {
+static void update_opcode_leave(zend_execute_data *execute_data) {
     DD_PRINTF("Update opcode leave");
     EX(call) = EX(call)->prev_execute_data;
 }
+
 static void execute_fcall(ddtrace_dispatch_t *dispatch, zval *this, zend_execute_data *execute_data,
-                          zval **return_value_ptr TSRMLS_DC) {
+                          zval **return_value_ptr) {
     zend_fcall_info fci = {0};
     zend_fcall_info_cache fcc = {0};
     char *error = NULL;
@@ -198,18 +317,18 @@ static void execute_fcall(ddtrace_dispatch_t *dispatch, zval *this, zend_execute
     zend_string *func_name = zend_string_init(ZEND_STRL(DDTRACE_CALLBACK_NAME), 0);
     func = EX(func);
     zend_create_closure(&closure, (zend_function *)zend_get_closure_method_def(&dispatch->callable),
-                        executed_method_class, executed_method_class, this TSRMLS_CC);
-    if (zend_fcall_info_init(&closure, 0, &fci, &fcc, NULL, &error TSRMLS_CC) != SUCCESS) {
+                        executed_method_class, executed_method_class, this);
+    if (zend_fcall_info_init(&closure, 0, &fci, &fcc, NULL, &error) != SUCCESS) {
         if (DDTRACE_G(strict_mode)) {
             const char *scope_name, *function_name;
 
             scope_name = (func->common.scope) ? ZSTR_VAL(func->common.scope->name) : NULL;
             function_name = ZSTR_VAL(func->common.function_name);
             if (scope_name) {
-                zend_throw_exception_ex(spl_ce_InvalidArgumentException, 0 TSRMLS_CC,
-                                        "cannot set override for %s::%s - %s", scope_name, function_name, error);
+                zend_throw_exception_ex(spl_ce_InvalidArgumentException, 0, "cannot set override for %s::%s - %s",
+                                        scope_name, function_name, error);
             } else {
-                zend_throw_exception_ex(spl_ce_InvalidArgumentException, 0 TSRMLS_CC, "cannot set override for %s - %s",
+                zend_throw_exception_ex(spl_ce_InvalidArgumentException, 0, "cannot set override for %s - %s",
                                         function_name, error);
             }
         }
@@ -220,7 +339,7 @@ static void execute_fcall(ddtrace_dispatch_t *dispatch, zval *this, zend_execute
         goto _exit_cleanup;
     }
 
-    ddtrace_setup_fcall(execute_data, &fci, return_value_ptr TSRMLS_CC);
+    ddtrace_setup_fcall(execute_data, &fci, return_value_ptr);
 
     // Move this to closure zval before zend_fcall_info_init()
     fcc.function_handler->common.function_name = func_name;
@@ -232,7 +351,7 @@ static void execute_fcall(ddtrace_dispatch_t *dispatch, zval *this, zend_execute
     zend_execute_data *prev_original_execute_data = DDTRACE_G(original_context).execute_data;
     DDTRACE_G(original_context).execute_data = execute_data;
 
-    zend_call_function(&fci, &fcc TSRMLS_CC);
+    zend_call_function(&fci, &fcc);
 
     DDTRACE_G(original_context).execute_data = prev_original_execute_data;
 
@@ -252,7 +371,7 @@ _exit_cleanup:
     DDTRACE_G(original_context).fbc = current_fbc;
 }
 
-static zend_always_inline void wrap_and_run(zend_execute_data *execute_data, ddtrace_dispatch_t *dispatch TSRMLS_DC) {
+static void wrap_and_run(zend_execute_data *execute_data, ddtrace_dispatch_t *dispatch) {
     zval *this = ddtrace_this(execute_data);
     const zend_op *opline = EX(opline);
 
@@ -260,84 +379,14 @@ static zend_always_inline void wrap_and_run(zend_execute_data *execute_data, ddt
     INIT_ZVAL(rv);
 
     zval *return_value = (RETURN_VALUE_USED(opline) ? EX_VAR(EX(opline)->result.var) : &rv);
-    execute_fcall(dispatch, this, EX(call), &return_value TSRMLS_CC);
+    execute_fcall(dispatch, this, EX(call), &return_value);
 
     if (!RETURN_VALUE_USED(opline)) {
         zval_dtor(&rv);
     }
 }
 
-int ddtrace_wrap_fcall(zend_execute_data *execute_data TSRMLS_DC) {
-    zend_function *current_fbc = NULL;
-    ddtrace_dispatch_t *dispatch = NULL;
-    if (!ddtrace_should_trace_call(execute_data, &current_fbc, &dispatch TSRMLS_CC)) {
-        return ddtrace_opcode_default_dispatch(execute_data TSRMLS_CC);
-    }
-    ddtrace_class_lookup_acquire(dispatch);  // protecting against dispatch being freed during php code execution
-    dispatch->busy = 1;                      // guard against recursion, catching only topmost execution
-
-    if (dispatch->run_as_postprocess) {
-        ddtrace_trace_dispatch(dispatch, current_fbc, execute_data TSRMLS_CC);
-    } else {
-        // Store original context for forwarding the call from userland
-        zend_function *previous_fbc = DDTRACE_G(original_context).fbc;
-        DDTRACE_G(original_context).fbc = current_fbc;
-        zend_function *previous_calling_fbc = DDTRACE_G(original_context).calling_fbc;
-
-        DDTRACE_G(original_context).calling_fbc = current_fbc->common.scope ? current_fbc : execute_data->func;
-
-        zval *this = ddtrace_this(execute_data);
-
-        zend_object *previous_this = DDTRACE_G(original_context).this;
-        DDTRACE_G(original_context).this = this ? Z_OBJ_P(this) : NULL;
-        zend_class_entry *previous_calling_ce = DDTRACE_G(original_context).calling_ce;
-
-        if (DDTRACE_G(original_context).this) {
-            GC_ADDREF(DDTRACE_G(original_context).this);
-        }
-        DDTRACE_G(original_context).calling_ce = Z_OBJ(execute_data->This) ? Z_OBJ(execute_data->This)->ce : NULL;
-
-        wrap_and_run(execute_data, dispatch TSRMLS_CC);
-        if (DDTRACE_G(original_context).this) {
-            GC_DELREF(DDTRACE_G(original_context).this);
-        }
-
-        // Restore original context
-        DDTRACE_G(original_context).calling_ce = previous_calling_ce;
-        DDTRACE_G(original_context).this = previous_this;
-        DDTRACE_G(original_context).calling_fbc = previous_calling_fbc;
-        DDTRACE_G(original_context).fbc = previous_fbc;
-
-        update_opcode_leave(execute_data TSRMLS_CC);
-    }
-
-    dispatch->busy = 0;
-    ddtrace_class_lookup_release(dispatch);
-
-    EX(opline)++;
-
-    return ZEND_USER_OPCODE_LEAVE;
-}
-
-void ddtrace_opcode_minit(void) {
-    _prev_icall_handler = zend_get_user_opcode_handler(ZEND_DO_ICALL);
-    _prev_ucall_handler = zend_get_user_opcode_handler(ZEND_DO_UCALL);
-    _prev_fcall_handler = zend_get_user_opcode_handler(ZEND_DO_FCALL);
-    _prev_fcall_by_name_handler = zend_get_user_opcode_handler(ZEND_DO_FCALL_BY_NAME);
-    zend_set_user_opcode_handler(ZEND_DO_ICALL, ddtrace_wrap_fcall);
-    zend_set_user_opcode_handler(ZEND_DO_UCALL, ddtrace_wrap_fcall);
-    zend_set_user_opcode_handler(ZEND_DO_FCALL, ddtrace_wrap_fcall);
-    zend_set_user_opcode_handler(ZEND_DO_FCALL_BY_NAME, ddtrace_wrap_fcall);
-}
-
-void ddtrace_opcode_mshutdown(void) {
-    zend_set_user_opcode_handler(ZEND_DO_ICALL, NULL);
-    zend_set_user_opcode_handler(ZEND_DO_UCALL, NULL);
-    zend_set_user_opcode_handler(ZEND_DO_FCALL, NULL);
-    zend_set_user_opcode_handler(ZEND_DO_FCALL_BY_NAME, NULL);
-}
-
-int ddtrace_opcode_default_dispatch(zend_execute_data *execute_data) {
+static int ddtrace_opcode_default_dispatch(zend_execute_data *execute_data) {
     if (!EX(opline)->opcode) {
         return ZEND_USER_OPCODE_DISPATCH;
     }
@@ -366,4 +415,74 @@ int ddtrace_opcode_default_dispatch(zend_execute_data *execute_data) {
             break;
     }
     return ZEND_USER_OPCODE_DISPATCH;
+}
+
+int ddtrace_wrap_fcall(zend_execute_data *execute_data) {
+    zend_function *current_fbc = NULL;
+    ddtrace_dispatch_t *dispatch = NULL;
+    if (!ddtrace_should_trace_call(execute_data, &current_fbc, &dispatch)) {
+        return ddtrace_opcode_default_dispatch(execute_data);
+    }
+    ddtrace_class_lookup_acquire(dispatch);  // protecting against dispatch being freed during php code execution
+    dispatch->busy = 1;                      // guard against recursion, catching only topmost execution
+
+    if (dispatch->run_as_postprocess) {
+        ddtrace_trace_dispatch(dispatch, current_fbc, execute_data);
+    } else {
+        // Store original context for forwarding the call from userland
+        zend_function *previous_fbc = DDTRACE_G(original_context).fbc;
+        DDTRACE_G(original_context).fbc = current_fbc;
+        zend_function *previous_calling_fbc = DDTRACE_G(original_context).calling_fbc;
+
+        DDTRACE_G(original_context).calling_fbc = current_fbc->common.scope ? current_fbc : execute_data->func;
+
+        zval *this = ddtrace_this(execute_data);
+
+        zend_object *previous_this = DDTRACE_G(original_context).this;
+        DDTRACE_G(original_context).this = this ? Z_OBJ_P(this) : NULL;
+        zend_class_entry *previous_calling_ce = DDTRACE_G(original_context).calling_ce;
+
+        if (DDTRACE_G(original_context).this) {
+            GC_ADDREF(DDTRACE_G(original_context).this);
+        }
+        DDTRACE_G(original_context).calling_ce = Z_OBJ(execute_data->This) ? Z_OBJ(execute_data->This)->ce : NULL;
+
+        wrap_and_run(execute_data, dispatch);
+        if (DDTRACE_G(original_context).this) {
+            GC_DELREF(DDTRACE_G(original_context).this);
+        }
+
+        // Restore original context
+        DDTRACE_G(original_context).calling_ce = previous_calling_ce;
+        DDTRACE_G(original_context).this = previous_this;
+        DDTRACE_G(original_context).calling_fbc = previous_calling_fbc;
+        DDTRACE_G(original_context).fbc = previous_fbc;
+
+        update_opcode_leave(execute_data);
+    }
+
+    dispatch->busy = 0;
+    ddtrace_class_lookup_release(dispatch);
+
+    EX(opline)++;
+
+    return ZEND_USER_OPCODE_LEAVE;
+}
+
+void ddtrace_opcode_minit(void) {
+    _prev_icall_handler = zend_get_user_opcode_handler(ZEND_DO_ICALL);
+    _prev_ucall_handler = zend_get_user_opcode_handler(ZEND_DO_UCALL);
+    _prev_fcall_handler = zend_get_user_opcode_handler(ZEND_DO_FCALL);
+    _prev_fcall_by_name_handler = zend_get_user_opcode_handler(ZEND_DO_FCALL_BY_NAME);
+    zend_set_user_opcode_handler(ZEND_DO_ICALL, ddtrace_wrap_fcall);
+    zend_set_user_opcode_handler(ZEND_DO_UCALL, ddtrace_wrap_fcall);
+    zend_set_user_opcode_handler(ZEND_DO_FCALL, ddtrace_wrap_fcall);
+    zend_set_user_opcode_handler(ZEND_DO_FCALL_BY_NAME, ddtrace_wrap_fcall);
+}
+
+void ddtrace_opcode_mshutdown(void) {
+    zend_set_user_opcode_handler(ZEND_DO_ICALL, NULL);
+    zend_set_user_opcode_handler(ZEND_DO_UCALL, NULL);
+    zend_set_user_opcode_handler(ZEND_DO_FCALL, NULL);
+    zend_set_user_opcode_handler(ZEND_DO_FCALL_BY_NAME, NULL);
 }
